actualizar

+++ spring-backend-apirest/src/main/java/com/example/spring/backend/apirest/models/repository/IClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.spring.backend.apirest.models.entity.Cliente;
 
-public interface IClienteRepository extends CrudRepository<Cliente,Long> {
+public interface IClienteRepository extends CrudRepository<Cliente,Integer> {
 
 }
